v1.1

+++ modules/diffSolvers.py
@@ -8,7 +8,6 @@ E-Mail  : bulmust@gmail.com
 import time
 import datetime
 import os as osCommand
-from distutils.dir_util import copy_tree
 from shutil import rmtree
 import numpy as np
 import RHS as func
@@ -21,7 +20,7 @@ import odeintw as integralSolver
 # ============================
 # Colors For Printing
 class tLog:
-    STATUS  = '\033[94m'+'[STATUS]    '+'\033[0m'
+    INFO  = '\033[94m'+'[INFO]    '+'\033[0m'
     OK      = '\033[92m'+'[OK]        '+'\033[0m'
     WARNING = '\033[93m'+'[WARNING]   '+'\033[0m'
     ERROR   = '\033[91m'+'[ERROR]     '+'\033[0m'
@@ -55,14 +54,15 @@ def odeintwSolver(init, RESULTS_SIMULATION_DIR):
     else:
         print(tLog.ERROR+"Data Folder Already Exists. Check The Folder: "+ DATA_FOLDER)
         print(tLog.EXIT)
-    # Create intermediateData folder
-    INTERMEDIATE_DATA_FOLDER= RESULTS_SIMULATION_DIR+ 'data/intermediateData/'
-    if not osCommand.path.exists(INTERMEDIATE_DATA_FOLDER):
-        osCommand.makedirs(INTERMEDIATE_DATA_FOLDER)
-        print(tLog.OK+"Intermediate Data Folder Created.")
-    else:
-        print(tLog.ERROR+"Intermediate Data Folder Already Exists. Check The Folder: "+ INTERMEDIATE_DATA_FOLDER)
-        print(tLog.EXIT)
+    if init.technicalParametersDic['holdIntermediateData']:
+        # Create intermediateData folder
+        INTERMEDIATE_DATA_FOLDER= RESULTS_SIMULATION_DIR+ 'data/intermediateData/'
+        if not osCommand.path.exists(INTERMEDIATE_DATA_FOLDER):
+            osCommand.makedirs(INTERMEDIATE_DATA_FOLDER)
+            print(tLog.OK+"Intermediate Data Folder Created.")
+        else:
+            print(tLog.ERROR+"Intermediate Data Folder Already Exists. Check The Folder: "+ INTERMEDIATE_DATA_FOLDER)
+            print(tLog.EXIT)
     # Change directory to data 
     # Code saves intermediate data
     osCommand.chdir(DATA_FOLDER)
@@ -75,71 +75,138 @@ def odeintwSolver(init, RESULTS_SIMULATION_DIR):
     # ============================
     # Solve E.o.M.
     # Starting time
-    print(tLog.STATUS+'Start solving E.o.M with odeintw. Current date and time :', timeInitialDate.strftime("%Y-%m-%d %H:%M:%S"))
-    if init.dim_rho_2totFlav_Bool:
-        # Default error tolerances
-        if (init.technicalParametersDic['tolerance_relativeError'] == 0 and init.technicalParametersDic['tolerance_absoluteError'] == 0):
-            if init.totFlav == 2:
-                # Two Flavor
-                rhoFinalAll,infow = integralSolver.odeintw(func.rhs2Flav_bigRho, init.rhoInit_flav, init.distAll_km, full_output=True, mxstep=500000000)
-            elif init.totFlav == 3:
-                # Three Flavor
-                rhoFinalAll,infow = integralSolver.odeintw(func.rhs3Flav_bigRho, init.rhoInit_flav, init.distAll_km, full_output=True, mxstep=500000000)
-            elif init.totFlav == 4:
-                # Four Flavor
-                rhoFinalAll,infow = integralSolver.odeintw(func.rhs4Flav_bigRho, init.rhoInit_flav, init.distAll_km, full_output=True, mxstep=500000000)
-        # User defined error tolerances
+    print(tLog.INFO+'Start solving E.o.M with odeintw. Current date and time :', timeInitialDate.strftime("%Y-%m-%d %H:%M:%S"))
+    # Hold intermediate data
+    if init.technicalParametersDic['holdIntermediateData']:
+        if init.dim_rho_2totFlav_Bool:
+            # Default error tolerances
+            if (init.technicalParametersDic['tolerance_relativeError'] == 0 and init.technicalParametersDic['tolerance_absoluteError'] == 0):
+                if init.totFlav == 2:
+                    # Two Flavor
+                    rhoFinalAll,infow = integralSolver.odeintw(func.rhs2Flav_bigRho, init.rhoInit_flav, init.distAll_km, full_output=True, mxstep=500000000)
+                elif init.totFlav == 3:
+                    # Three Flavor
+                    rhoFinalAll,infow = integralSolver.odeintw(func.rhs3Flav_bigRho, init.rhoInit_flav, init.distAll_km, full_output=True, mxstep=500000000)
+                elif init.totFlav == 4:
+                    # Four Flavor
+                    rhoFinalAll,infow = integralSolver.odeintw(func.rhs4Flav_bigRho, init.rhoInit_flav, init.distAll_km, full_output=True, mxstep=500000000)
+            # User defined error tolerances
+            else:
+                if init.totFlav == 2:
+                    # Two Flavor
+                    rhoFinalAll,infow = integralSolver.odeintw(func.rhs2Flav_bigRho, init.rhoInit_flav\
+                        , init.distAll_km, full_output=True, rtol=init.technicalParametersDic['tolerance_relativeError']\
+                        , atol=init.technicalParametersDic['tolerance_absoluteError']\
+                        , mxstep=500000000)
+                elif init.totFlav == 3:
+                    # Three Flavor
+                    rhoFinalAll,infow = integralSolver.odeintw(func.rhs3Flav_bigRho, init.rhoInit_flav\
+                        , init.distAll_km, full_output=True, rtol=init.technicalParametersDic['tolerance_relativeError']\
+                        , atol=init.technicalParametersDic['tolerance_absoluteError']\
+                        , mxstep=500000000)
+                elif init.totFlav == 4:
+                    # Four Flavor
+                    rhoFinalAll,infow = integralSolver.odeintw(func.rhs4Flav_bigRho, init.rhoInit_flav\
+                        , init.distAll_km, full_output=True, rtol=init.technicalParametersDic['tolerance_relativeError']\
+                        , atol=init.technicalParametersDic['tolerance_absoluteError']\
+                        , mxstep=500000000)
         else:
-            if init.totFlav == 2:
-                # Two Flavor
-                rhoFinalAll,infow = integralSolver.odeintw(func.rhs2Flav_bigRho, init.rhoInit_flav\
-                    , init.distAll_km, full_output=True, rtol=init.technicalParametersDic['tolerance_relativeError']\
-                    , atol=init.technicalParametersDic['tolerance_absoluteError']\
-                    , mxstep=500000000)
-            elif init.totFlav == 3:
-                # Three Flavor
-                rhoFinalAll,infow = integralSolver.odeintw(func.rhs3Flav_bigRho, init.rhoInit_flav\
-                    , init.distAll_km, full_output=True, rtol=init.technicalParametersDic['tolerance_relativeError']\
-                    , atol=init.technicalParametersDic['tolerance_absoluteError']\
-                    , mxstep=500000000)
-            elif init.totFlav == 4:
-                # Four Flavor
-                rhoFinalAll,infow = integralSolver.odeintw(func.rhs4Flav_bigRho, init.rhoInit_flav\
-                    , init.distAll_km, full_output=True, rtol=init.technicalParametersDic['tolerance_relativeError']\
-                    , atol=init.technicalParametersDic['tolerance_absoluteError']\
-                    , mxstep=500000000)
+            # Default error tolerances
+            if (init.technicalParametersDic['tolerance_relativeError'] == 0 and init.technicalParametersDic['tolerance_absoluteError'] == 0):
+                if init.totFlav == 2:
+                    # Two Flavor
+                    rhoFinalAll,infow = integralSolver.odeintw(func.rhs2Flav, init.rhoInit_flav, init.distAll_km, full_output=True, mxstep=500000000)
+                elif init.totFlav == 3:
+                    # Three Flavor
+                    rhoFinalAll,infow = integralSolver.odeintw(func.rhs3Flav, init.rhoInit_flav, init.distAll_km, full_output=True, mxstep=500000000)
+                elif init.totFlav == 4:
+                    # Four Flavor
+                    rhoFinalAll,infow = integralSolver.odeintw(func.rhs4Flav, init.rhoInit_flav, init.distAll_km, full_output=True, mxstep=500000000)    
+            # User defined error tolerances
+            else:
+                if init.totFlav == 2:
+                    # Two Flavor
+                    rhoFinalAll,infow = integralSolver.odeintw(func.rhs2Flav, init.rhoInit_flav\
+                        , init.distAll_km, full_output=True, rtol=init.technicalParametersDic['tolerance_relativeError']\
+                        , atol=init.technicalParametersDic['tolerance_absoluteError']\
+                        , mxstep=500000000)
+                elif init.totFlav == 3:
+                    # Three Flavor
+                    rhoFinalAll,infow = integralSolver.odeintw(func.rhs3Flav, init.rhoInit_flav\
+                        , init.distAll_km, full_output=True, rtol=init.technicalParametersDic['tolerance_relativeError']\
+                        , atol=init.technicalParametersDic['tolerance_absoluteError']\
+                        , mxstep=500000000)
+                elif init.totFlav == 4:
+                    # Four Flavor
+                    rhoFinalAll,infow = integralSolver.odeintw(func.rhs4Flav, init.rhoInit_flav\
+                        , init.distAll_km, full_output=True, rtol=init.technicalParametersDic['tolerance_relativeError']\
+                        , atol=init.technicalParametersDic['tolerance_absoluteError']\
+                        , mxstep=500000000)
     else:
-        # Default error tolerances
-        if (init.technicalParametersDic['tolerance_relativeError'] == 0 and init.technicalParametersDic['tolerance_absoluteError'] == 0):
-            if init.totFlav == 2:
-                # Two Flavor
-                rhoFinalAll,infow = integralSolver.odeintw(func.rhs2Flav, init.rhoInit_flav, init.distAll_km, full_output=True, mxstep=500000000)
-            elif init.totFlav == 3:
-                # Three Flavor
-                rhoFinalAll,infow = integralSolver.odeintw(func.rhs3Flav, init.rhoInit_flav, init.distAll_km, full_output=True, mxstep=500000000)
-            elif init.totFlav == 4:
-                # Four Flavor
-                rhoFinalAll,infow = integralSolver.odeintw(func.rhs4Flav, init.rhoInit_flav, init.distAll_km, full_output=True, mxstep=500000000)    
-        # User defined error tolerances
+        if init.dim_rho_2totFlav_Bool:
+            # Default error tolerances
+            if (init.technicalParametersDic['tolerance_relativeError'] == 0 and init.technicalParametersDic['tolerance_absoluteError'] == 0):
+                if init.totFlav == 2:
+                    # Two Flavor
+                    rhoFinalAll,infow = integralSolver.odeintw(func.rhs2Flav_bigRho_noInterData, init.rhoInit_flav, init.distAll_km, full_output=True, mxstep=500000000)
+                elif init.totFlav == 3:
+                    # Three Flavor
+                    rhoFinalAll,infow = integralSolver.odeintw(func.rhs3Flav_bigRho_noInterData, init.rhoInit_flav, init.distAll_km, full_output=True, mxstep=500000000)
+                elif init.totFlav == 4:
+                    # Four Flavor
+                    rhoFinalAll,infow = integralSolver.odeintw(func.rhs4Flav_bigRho_noInterData, init.rhoInit_flav, init.distAll_km, full_output=True, mxstep=500000000)
+            # User defined error tolerances
+            else:
+                if init.totFlav == 2:
+                    # Two Flavor
+                    rhoFinalAll,infow = integralSolver.odeintw(func.rhs2Flav_bigRho_noInterData, init.rhoInit_flav\
+                        , init.distAll_km, full_output=True, rtol=init.technicalParametersDic['tolerance_relativeError']\
+                        , atol=init.technicalParametersDic['tolerance_absoluteError']\
+                        , mxstep=500000000)
+                elif init.totFlav == 3:
+                    # Three Flavor
+                    rhoFinalAll,infow = integralSolver.odeintw(func.rhs3Flav_bigRho_noInterData, init.rhoInit_flav\
+                        , init.distAll_km, full_output=True, rtol=init.technicalParametersDic['tolerance_relativeError']\
+                        , atol=init.technicalParametersDic['tolerance_absoluteError']\
+                        , mxstep=500000000)
+                elif init.totFlav == 4:
+                    # Four Flavor
+                    rhoFinalAll,infow = integralSolver.odeintw(func.rhs4Flav_bigRho_noInterData, init.rhoInit_flav\
+                        , init.distAll_km, full_output=True, rtol=init.technicalParametersDic['tolerance_relativeError']\
+                        , atol=init.technicalParametersDic['tolerance_absoluteError']\
+                        , mxstep=500000000)
         else:
-            if init.totFlav == 2:
-                # Two Flavor
-                rhoFinalAll,infow = integralSolver.odeintw(func.rhs2Flav, init.rhoInit_flav\
-                    , init.distAll_km, full_output=True, rtol=init.technicalParametersDic['tolerance_relativeError']\
-                    , atol=init.technicalParametersDic['tolerance_absoluteError']\
-                    , mxstep=500000000)
-            elif init.totFlav == 3:
-                # Three Flavor
-                rhoFinalAll,infow = integralSolver.odeintw(func.rhs3Flav, init.rhoInit_flav\
-                    , init.distAll_km, full_output=True, rtol=init.technicalParametersDic['tolerance_relativeError']\
-                    , atol=init.technicalParametersDic['tolerance_absoluteError']\
-                    , mxstep=500000000)
-            elif init.totFlav == 4:
-                # Four Flavor
-                rhoFinalAll,infow = integralSolver.odeintw(func.rhs4Flav, init.rhoInit_flav\
-                    , init.distAll_km, full_output=True, rtol=init.technicalParametersDic['tolerance_relativeError']\
-                    , atol=init.technicalParametersDic['tolerance_absoluteError']\
-                    , mxstep=500000000)
+            # Default error tolerances
+            if (init.technicalParametersDic['tolerance_relativeError'] == 0 and init.technicalParametersDic['tolerance_absoluteError'] == 0):
+                if init.totFlav == 2:
+                    # Two Flavor
+                    rhoFinalAll,infow = integralSolver.odeintw(func.rhs2Flav_noInterData, init.rhoInit_flav, init.distAll_km, full_output=True, mxstep=500000000)
+                elif init.totFlav == 3:
+                    # Three Flavor
+                    rhoFinalAll,infow = integralSolver.odeintw(func.rhs3Flav_noInterData, init.rhoInit_flav, init.distAll_km, full_output=True, mxstep=500000000)
+                elif init.totFlav == 4:
+                    # Four Flavor
+                    rhoFinalAll,infow = integralSolver.odeintw(func.rhs4Flav_noInterData, init.rhoInit_flav, init.distAll_km, full_output=True, mxstep=500000000)    
+            # User defined error tolerances
+            else:
+                if init.totFlav == 2:
+                    # Two Flavor
+                    rhoFinalAll,infow = integralSolver.odeintw(func.rhs2Flav_noInterData, init.rhoInit_flav\
+                        , init.distAll_km, full_output=True, rtol=init.technicalParametersDic['tolerance_relativeError']\
+                        , atol=init.technicalParametersDic['tolerance_absoluteError']\
+                        , mxstep=500000000)
+                elif init.totFlav == 3:
+                    # Three Flavor
+                    rhoFinalAll,infow = integralSolver.odeintw(func.rhs3Flav_noInterData, init.rhoInit_flav\
+                        , init.distAll_km, full_output=True, rtol=init.technicalParametersDic['tolerance_relativeError']\
+                        , atol=init.technicalParametersDic['tolerance_absoluteError']\
+                        , mxstep=500000000)
+                elif init.totFlav == 4:
+                    # Four Flavor
+                    rhoFinalAll,infow = integralSolver.odeintw(func.rhs4Flav_noInterData, init.rhoInit_flav\
+                        , init.distAll_km, full_output=True, rtol=init.technicalParametersDic['tolerance_relativeError']\
+                        , atol=init.technicalParametersDic['tolerance_absoluteError']\
+                        , mxstep=500000000)
     # ============================  
     
     # ============================  
@@ -161,25 +228,25 @@ def odeintwSolver(init, RESULTS_SIMULATION_DIR):
     minutes = tmpTimeCalculation // 60
     tmpTimeCalculation %= 60
     seconds = tmpTimeCalculation
-    print(tLog.STATUS+"Total Process Time, %6.2fs,  d:h:m:s : %d:%d:%d:%d"% (tmpTimeCalculationSec, days, hours, minutes, seconds))
+    print(tLog.INFO+"Total Process Time, %6.2fs,  d:h:m:s : %d:%d:%d:%d"% (tmpTimeCalculationSec, days, hours, minutes, seconds))
     # ============================
     # Save Total Process Time to file
-    # Do not color the [STATUS]
+    # Do not color the [INFO]
     fileObject = open('distanceAndTime.log','a+')
     print('Total Process Time, %6.2fs,  d:h:m:s : %d:%d:%d:%d'% (tmpTimeCalculationSec, days, hours, minutes, seconds), file=fileObject)
     fileObject.close()
     # ============================
     # Number Of Evolution
-    print(tLog.STATUS+'Number of evaluation: %d' % infow['nfe'][-1])
+    print(tLog.INFO+'Number of evaluation: %d' % infow['nfe'][-1])
     # ============================
     # Check differential equation solver method changed or not
     methodIndicators = infow['mused']
     if methodIndicators[0] == 1:
-        print(tLog.STATUS+'Differential equation solver started with adams (nonstiff) method.')
+        print(tLog.INFO+'Differential equation solver started with adams (nonstiff) method.')
     else:
-        print(tLog.STATUS+'Differential equation solver started with bdf (stiff) method.')
-    indicatorChangePrint = [tLog.STATUS+'Differential equation solver method is adams (nonstiff)'\
-        , tLog.STATUS+'Differential equation solver method is bdf   (stiff)   ']
+        print(tLog.INFO+'Differential equation solver started with bdf (stiff) method.')
+    indicatorChangePrint = [tLog.INFO+'Differential equation solver method is adams (nonstiff)'\
+        , tLog.INFO+'Differential equation solver method is bdf   (stiff)   ']
     # ============================
     # If tmpIndicatorCheckControl == 1, the method changed.
     tmpIndicatorCheckControl = 0
@@ -204,9 +271,9 @@ def odeintwSolver(init, RESULTS_SIMULATION_DIR):
     # ============================
     # Print if the method is changed or not
     if tmpIndicatorCheckControl == 0:
-        print(tLog.STATUS+'Differential equation solver method have not changed.')
+        print(tLog.INFO+'Differential equation solver method have not changed.')
     elif tmpIndicatorCheckControl == 1:
-        print(tLog.STATUS+'Differential equation solver method have changed.')
+        print(tLog.INFO+'Differential equation solver method have changed.')
     # ============================
     '''
     ALL INFOS
@@ -229,8 +296,9 @@ def odeintwSolver(init, RESULTS_SIMULATION_DIR):
         1: adams (nonstiff), 2: bdf (stiff)
     '''
     # ============================
-    # Remove intermediateData folder
-    rmtree(INTERMEDIATE_DATA_FOLDER)
+    if init.technicalParametersDic['holdIntermediateData']:
+        # Remove intermediateData folder
+        rmtree(INTERMEDIATE_DATA_FOLDER)
     # Return Back To RESULTS_SIMULATION_DIR Folder
     osCommand.chdir(RESULTS_SIMULATION_DIR)
